Reduce logging Improve error handling Rename class as requested Added license header Improve url parsing added robots.txt
